rpi/wildlife_photographer/requirements.txt
https://github.com/google-coral/pycoral/releases/download/v1.0.1/tflite_runtime-2.5.0-cp37-cp37m-linux_armv7l.whl
picamera
Pillow
telegram-send
